🚚 [variables] Move module `prompts` to `adapters`
